requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1

# Scheduling
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# For Chinese text processing
chardet==5.2.0
